style : SVG 파일 에러 이슈 해결 rename : 코딩 컨벤션에 따른 폴더명 재구성
