+++ src/components/TripEdit/TripPlaneIcon.jsx
